Create README.md
This is an added optional extended description that I decided to put. Just to let everyone know that I added this optional extended description and it wasn't a bug. This added optional extended description was added by me and this is my added optional extended description. If you want my added optional extended description than too bad. This is my added optional extended description and you can't have it. If you want my added optional extended description you're gonna have to go through me first. I would do anything to protect my added optional extended description. if I get married in the future than me and my partner may decide that we are now sharing my added optional extended  description but we will discuss that when the time comes. In the  event of divorce, my then ex-wife may decide that she wants my added optional extended description at which point it would become hers, and would no longer be my property. In the event of my untimely or timely death, I will pass on this here added  optional extended description to my next of kin. In the event that this added optional extended description is to change hands, please treat the new owner with utter respect as it would be an  honor for this added optional extended description to outlive me.